# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.Activity:
    properties:
      activityname:
        type: string
      added:
        type: string
      agency:
        type: integer
      hours:
        type: integer
      place:
        type: integer
      student:
        type: integer
      year:
        type: integer
    type: object
...
  ent.Activity:
    properties:
      ACTIVITYNAME:
        description: ACTIVITYNAME holds the value of the "ACTIVITYNAME" field.
        type: string
      ADDED:
        description: ADDED holds the value of the "ADDED" field.
        type: string
      HOURS:
        description: HOURS holds the value of the "HOURS" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.ActivityEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ActivityQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.ActivityEdges:
    properties:
      actiAgen:
        $ref: '#/definitions/ent.Agency'
        description: ActiAgen holds the value of the acti_agen edge.
        type: object
      actiPlace:
        $ref: '#/definitions/ent.Place'
        description: ActiPlace holds the value of the acti_place edge.
        type: object
      actiStud:
        $ref: '#/definitions/ent.Student'
        description: ActiStud holds the value of the acti_stud edge.
        type: object
      actiYear:
        $ref: '#/definitions/ent.Year'
        description: ActiYear holds the value of the acti_year edge.
        type: object
    type: object
  ent.Agency:
    properties:
      AGENCY:
        description: AGENCY holds the value of the "AGENCY" field.
        type: string
      edges:
        $ref: '#/definitions/ent.AgencyEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the AgencyQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.AgencyEdges:
    properties:
      agenActi:
        description: AgenActi holds the value of the agen_acti edge.
        items:
          $ref: '#/definitions/ent.Activity'
        type: array
    type: object
  ent.Continent:
    properties:
      continent:
        description: Continent holds the value of the "continent" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ContinentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ContinentQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.ContinentEdges:
    properties:
      contProv:
        description: ContProv holds the value of the cont_prov edge.
        items:
          $ref: '#/definitions/ent.Province'
        type: array
    type: object
  ent.Country:
    properties:
      country:
        description: Country holds the value of the "country" field.
        type: string
      edges:
        $ref: '#/definitions/ent.CountryEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CountryQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.CountryEdges:
    properties:
      counProv:
        description: CounProv holds the value of the coun_prov edge.
        items:
          $ref: '#/definitions/ent.Province'
        type: array
    type: object
  ent.Course:
    properties:
      course:
        description: Course holds the value of the "course" field.
        type: string
      edges:
        $ref: '#/definitions/ent.CourseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the CourseQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.CourseEdges:
    properties:
      courDegr:
        $ref: '#/definitions/ent.Degree'
        description: CourDegr holds the value of the cour_degr edge.
        type: object
      courFacu:
        $ref: '#/definitions/ent.Faculty'
        description: CourFacu holds the value of the cour_facu edge.
        type: object
      courInst:
        $ref: '#/definitions/ent.Institution'
        description: CourInst holds the value of the cour_inst edge.
        type: object
    type: object
  ent.Degree:
    properties:
      degree:
        description: Degree holds the value of the "degree" field.
        type: string
      edges:
        $ref: '#/definitions/ent.DegreeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DegreeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.DegreeEdges:
    properties:
      degrCour:
        description: DegrCour holds the value of the degr_cour edge.
        items:
          $ref: '#/definitions/ent.Course'
        type: array
      degrStud:
        description: DegrStud holds the value of the degr_stud edge.
        items:
          $ref: '#/definitions/ent.Student'
        type: array
    type: object
  ent.Emp:
    properties:
      id:
        description: ID of the ent.
        type: integer
      pass:
        description: Pass holds the value of the "pass" field.
        type: string
      role:
        description: Role holds the value of the "role" field.
        type: string
      user:
        description: User holds the value of the "user" field.
        type: string
    type: object
  ent.Faculty:
    properties:
      edges:
        $ref: '#/definitions/ent.FacultyEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the FacultyQuery when eager-loading is set.
        type: object
      faculty:
        description: Faculty holds the value of the "faculty" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.FacultyEdges:
    properties:
      facuCour:
        description: FacuCour holds the value of the facu_cour edge.
        items:
          $ref: '#/definitions/ent.Course'
        type: array
      facuProf:
        description: FacuProf holds the value of the facu_prof edge.
        items:
          $ref: '#/definitions/ent.Professor'
        type: array
    type: object
  ent.Gender:
    properties:
      edges:
        $ref: '#/definitions/ent.GenderEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the GenderQuery when eager-loading is set.
        type: object
      gender:
        description: Gender holds the value of the "gender" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.GenderEdges:
    properties:
      gendStud:
        description: GendStud holds the value of the gend_stud edge.
        items:
          $ref: '#/definitions/ent.Student'
        type: array
    type: object
  ent.Institution:
    properties:
      edges:
        $ref: '#/definitions/ent.InstitutionEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the InstitutionQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      institution:
        description: Institution holds the value of the "institution" field.
        type: string
    type: object
  ent.InstitutionEdges:
    properties:
      instCour:
        description: InstCour holds the value of the inst_cour edge.
        items:
          $ref: '#/definitions/ent.Course'
        type: array
    type: object
  ent.Place:
    properties:
      PLACE:
        description: PLACE holds the value of the "PLACE" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PlaceEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PlaceQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.PlaceEdges:
    properties:
      placeActi:
        description: PlaceActi holds the value of the place_acti edge.
        items:
          $ref: '#/definitions/ent.Activity'
        type: array
    type: object
  ent.Prefix:
    properties:
      edges:
        $ref: '#/definitions/ent.PrefixEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PrefixQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      prefix:
        description: Prefix holds the value of the "prefix" field.
        type: string
    type: object
  ent.PrefixEdges:
    properties:
      preProf:
        description: PreProf holds the value of the pre_prof edge.
        items:
          $ref: '#/definitions/ent.Professor'
        type: array
      prefStud:
        description: PrefStud holds the value of the pref_stud edge.
        items:
          $ref: '#/definitions/ent.Student'
        type: array
    type: object
  ent.Professor:
    properties:
      edges:
        $ref: '#/definitions/ent.ProfessorEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ProfessorQuery when eager-loading is set.
        type: object
      email:
        description: Email holds the value of the "email" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      tel:
        description: Tel holds the value of the "tel" field.
        type: string
    type: object
  ent.ProfessorEdges:
    properties:
      profFacu:
        $ref: '#/definitions/ent.Faculty'
        description: ProfFacu holds the value of the prof_facu edge.
        type: object
      profPre:
        $ref: '#/definitions/ent.Prefix'
        description: ProfPre holds the value of the prof_pre edge.
        type: object
      profPros:
        $ref: '#/definitions/ent.Professorship'
        description: ProfPros holds the value of the prof_pros edge.
        type: object
    type: object
  ent.Professorship:
    properties:
      edges:
        $ref: '#/definitions/ent.ProfessorshipEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ProfessorshipQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      professorship:
        description: Professorship holds the value of the "professorship" field.
        type: string
    type: object
  ent.ProfessorshipEdges:
    properties:
      prosProf:
        description: ProsProf holds the value of the pros_prof edge.
        items:
          $ref: '#/definitions/ent.Professor'
        type: array
    type: object
  ent.Province:
    properties:
      district:
        description: District holds the value of the "district" field.
        type: string
      edges:
        $ref: '#/definitions/ent.ProvinceEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ProvinceQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      postal:
        description: Postal holds the value of the "postal" field.
        type: integer
      province:
        description: Province holds the value of the "province" field.
        type: string
      subdistrict:
        description: Subdistrict holds the value of the "subdistrict" field.
        type: string
    type: object
  ent.ProvinceEdges:
    properties:
      distStud:
        description: DistStud holds the value of the dist_stud edge.
        items:
          $ref: '#/definitions/ent.Student'
        type: array
      postStud:
        description: PostStud holds the value of the post_stud edge.
        items:
          $ref: '#/definitions/ent.Student'
        type: array
      provCont:
        $ref: '#/definitions/ent.Continent'
        description: ProvCont holds the value of the prov_cont edge.
        type: object
      provCoun:
        $ref: '#/definitions/ent.Country'
        description: ProvCoun holds the value of the prov_coun edge.
        type: object
      provRegi:
        $ref: '#/definitions/ent.Region'
        description: ProvRegi holds the value of the prov_regi edge.
        type: object
      provStud:
        description: ProvStud holds the value of the prov_stud edge.
        items:
          $ref: '#/definitions/ent.Student'
        type: array
      subdStud:
        description: SubdStud holds the value of the subd_stud edge.
        items:
          $ref: '#/definitions/ent.Student'
        type: array
    type: object
  ent.Region:
    properties:
      edges:
        $ref: '#/definitions/ent.RegionEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the RegionQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.RegionEdges:
    properties:
      regiProv:
        description: RegiProv holds the value of the regi_prov edge.
        items:
          $ref: '#/definitions/ent.Province'
        type: array
    type: object
  ent.Results:
    properties:
      edges:
        $ref: '#/definitions/ent.ResultsEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ResultsQuery when eager-loading is set.
        type: object
      grade:
        description: Grade holds the value of the "grade" field.
        type: number
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.ResultsEdges:
    properties:
      resuStud:
        $ref: '#/definitions/ent.Student'
        description: ResuStud holds the value of the resu_stud edge.
        type: object
      resuSubj:
        $ref: '#/definitions/ent.Subject'
        description: ResuSubj holds the value of the resu_subj edge.
        type: object
      resuTerm:
        $ref: '#/definitions/ent.Term'
        description: ResuTerm holds the value of the resu_term edge.
        type: object
      resuYear:
        $ref: '#/definitions/ent.Year'
        description: ResuYear holds the value of the resu_year edge.
        type: object
    type: object
  ent.Student:
    properties:
      edges:
        $ref: '#/definitions/ent.StudentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the StudentQuery when eager-loading is set.
        type: object
      email:
        description: Email holds the value of the "email" field.
        type: string
      fname:
        description: Fname holds the value of the "fname" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      lname:
        description: Lname holds the value of the "lname" field.
        type: string
      recent_address:
        description: RecentAddress holds the value of the "recent_address" field.
        type: string
      schoolname:
        description: Schoolname holds the value of the "schoolname" field.
        type: string
      telephone:
        description: Telephone holds the value of the "telephone" field.
        type: integer
    type: object
  ent.StudentEdges:
    properties:
      studActi:
        description: StudActi holds the value of the stud_acti edge.
        items:
          $ref: '#/definitions/ent.Activity'
        type: array
      studDegr:
        $ref: '#/definitions/ent.Degree'
        description: StudDegr holds the value of the stud_degr edge.
        type: object
      studDist:
        $ref: '#/definitions/ent.Province'
        description: StudDist holds the value of the stud_dist edge.
        type: object
      studGend:
        $ref: '#/definitions/ent.Gender'
        description: StudGend holds the value of the stud_gend edge.
        type: object
      studPost:
        $ref: '#/definitions/ent.Province'
        description: StudPost holds the value of the stud_post edge.
        type: object
      studPref:
        $ref: '#/definitions/ent.Prefix'
        description: StudPref holds the value of the stud_pref edge.
        type: object
      studProv:
        $ref: '#/definitions/ent.Province'
        description: StudProv holds the value of the stud_prov edge.
        type: object
      studResu:
        description: StudResu holds the value of the stud_resu edge.
        items:
          $ref: '#/definitions/ent.Results'
        type: array
      studSubd:
        $ref: '#/definitions/ent.Province'
        description: StudSubd holds the value of the stud_subd edge.
        type: object
    type: object
  ent.Subject:
    properties:
      code:
        description: Code holds the value of the "code" field.
        type: integer
      creditpiont:
        description: Creditpiont holds the value of the "creditpiont" field.
        type: integer
      edges:
        $ref: '#/definitions/ent.SubjectEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the SubjectQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      subjects:
        description: Subjects holds the value of the "subjects" field.
        type: string
    type: object
  ent.SubjectEdges:
    properties:
      subjResu:
        description: SubjResu holds the value of the subj_resu edge.
        items:
          $ref: '#/definitions/ent.Results'
        type: array
    type: object
  ent.Term:
    properties:
      edges:
        $ref: '#/definitions/ent.TermEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the TermQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      semester:
        description: Semester holds the value of the "semester" field.
        type: integer
    type: object
  ent.TermEdges:
    properties:
      termActi:
        description: TermActi holds the value of the term_acti edge.
        items:
          $ref: '#/definitions/ent.Activity'
        type: array
      termResu:
        description: TermResu holds the value of the term_resu edge.
        items:
          $ref: '#/definitions/ent.Results'
        type: array
    type: object
  ent.Year:
    properties:
      edges:
        $ref: '#/definitions/ent.YearEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the YearQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      years:
        description: Years holds the value of the "years" field.
        type: integer
    type: object
  ent.YearEdges:
    properties:
      yearActi:
        description: YearActi holds the value of the year_acti edge.
        items:
          $ref: '#/definitions/ent.Activity'
        type: array
      yearResu:
        description: YearResu holds the value of the year_resu edge.
        items:
          $ref: '#/definitions/ent.Results'
        type: array
      yearTerm:
        description: YearTerm holds the value of the year_term edge.
        items:
          $ref: '#/definitions/ent.Term'
        type: array
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT SA Example API Playlist Vidoe
  version: "1.0"
paths:
  /activitys:
    get:
      description: list activity entities
      operationId: list-activity
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Activity'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List activity entities
    post:
      consumes:
      - application/json
      description: Create activity
      operationId: create-activity
      parameters:
      - description: Activity entity
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/controllers.Activity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Activity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create activity
  /activitys/{id}:
    delete:
      description: get activity by ID
      operationId: delete-activity
      parameters:
      - description: Activity ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a activity entity by ID
...
  /emps:
    get:
      description: list emp entities
      operationId: list-emp
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Emp'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List emp entities
  /emps/{id}:
    get:
      description: get emp by ID
      operationId: get-emp
      parameters:
      - description: Emp ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Emp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a emp entity by ID
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    authorizationUrl: ""
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    authorizationUrl: ""
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"